+++ feature/profile/src/commonMain/kotlin/kpt/feature/profile/ProfileScreen.kt
@@ -9,74 +9,29 @@
  */
 package kpt.feature.profile
 
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
-import kpt.core.base.designsystem.component.HeroCard
-import kpt.core.base.ui.AppInfo
-import kpt.core.designsystem.icon.AppIcons
-import kpt.core.designsystem.theme.spacing
 import kpt.core.ui.scaffold.KptScaffold
-import kpt.feature.profile.generated.resources.Res
-import kpt.feature.profile.generated.resources.screens_profile_local_message
-import kpt.feature.profile.generated.resources.screens_profile_local_title
-import org.jetbrains.compose.resources.stringResource
 
+/**
+ * The profile tab shell — a framework-owned [KptScaffold] that carries ZERO demo/content imports and
+ * simply renders whatever opaque [profileBody] the caller supplies. The fork owns that body via
+ * `cmp-navigation`'s `BackboneRegistry.profileBody` (default: [kpt.feature.profile.demo.ProfileDemoBody]),
+ * mirroring the `home` shell/seam split — so a template sync full-copies this shell while the fork's
+ * profile inner content survives. (WS01 base-feature seam, epic AC7.)
+ */
 @Composable
-internal fun ProfileScreen(modifier: Modifier = Modifier) {
-    ProfileScreenContent(
+internal fun ProfileScreen(
+    profileBody: @Composable () -> Unit = {},
+    modifier: Modifier = Modifier,
+) {
+    KptScaffold(
         modifier = modifier
             .fillMaxSize()
             .testTag(TestTags.Profile.SCREEN),
-    )
-}
-
-@Composable
-internal fun ProfileScreenContent(modifier: Modifier = Modifier) {
-    val sp = MaterialTheme.spacing
-    KptScaffold(modifier = modifier) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(sp.lg),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Center,
-        ) {
-            HeroCard {
-                Column(
-                    modifier = Modifier.padding(sp.lg),
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    verticalArrangement = Arrangement.spacedBy(sp.sm),
-                ) {
-                    Icon(
-                        imageVector = AppIcons.AccountCircle,
-                        contentDescription = null,
-                        modifier = Modifier.size(64.dp),
-                    )
-                    Text(
-                        text = stringResource(Res.string.screens_profile_local_title),
-                        style = MaterialTheme.typography.titleMedium,
-                        fontWeight = FontWeight.SemiBold,
-                    )
-                    Text(
-                        // App display name templated in from the common AppInfo accessor, never hardcoded.
-                        text = stringResource(Res.string.screens_profile_local_message, AppInfo.appDisplayName),
-                        style = MaterialTheme.typography.bodyMedium,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    )
-                }
-            }
-        }
+    ) {
+        profileBody()
     }
 }
