+++ feature/settings/src/commonTest/kotlin/kpt/feature/settings/SettingsScreenUiTest.kt
@@ -14,16 +14,17 @@ import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.runComposeUiTest
 import kpt.core.designsystem.theme.KptTheme
+import kpt.feature.settings.demo.SettingsDemoBody
 import kotlin.test.Test
 
 /**
- * Compose Multiplatform UI test for [SettingsScreen].
+ * Compose Multiplatform UI test for the settings inner body [SettingsDemoBody] (the fork-owned seam
+ * rendered by `cmp-navigation`'s `BackboneRegistry.settingsBody` inside the template `SettingsScreen`
+ * shell).
  *
- * SettingsScreen is a pure-UI screen — all dialog state is managed with
- * [rememberSaveable] and [rememberAnalyticsHelper] internally; no ViewModel is
- * needed. The test renders it directly inside [KptTheme] with no-op callbacks
- * and asserts the always-present root scaffold node identified by
- * [TestTags.Settings.SCREEN].
+ * It is a pure-UI body — all dialog state is managed internally; no ViewModel is needed. The test
+ * renders it directly inside [KptTheme] with no-op callbacks and asserts the always-present root
+ * scaffold node identified by [TestTags.Settings.SCREEN].
  */
 @OptIn(ExperimentalTestApi::class)
 class SettingsScreenUiTest {
@@ -32,7 +33,8 @@ class SettingsScreenUiTest {
     fun screenIsDisplayed() = runComposeUiTest {
         setContent {
             KptTheme {
-                SettingsScreen(
+                // In production cmp-navigation's BackboneRegistry.settingsBody supplies this body.
+                SettingsDemoBody(
                     onBackClick = {},
                     onSyncAndDraftsClick = {},
                 )
